server/pyproject.toml
[project]
name = "enthusiast"
requires-python = ">=3.10, <4.0"
dependencies = [
    "celery (==5.5.3)",
    "channels (==4.2.2)",
    "channels-redis (==4.2.1)",
    "daphne (==4.2.1)",
    "django (==5.2.9)",
    "django-cors-headers (==4.7.0)",
    "django-environ (==0.12.0)",
    "djangorestframework (==3.16.0)",
    "drf-yasg (==1.21.10)",
    "enthusiast-common (==1.5.0)",
    "enthusiast-source-sample (==1.2.0)",
    "enthusiast-model-openai (==1.3.0)",
    "langchain (==0.3.27)",
    "pgvector (==0.4.1)",
    "psycopg2 (==2.9.10)",
    "pytest-django (==4.11.1)",
    "redis (==6.2.0)",
    "uvicorn (==0.30.6)",
    "whitenoise (==6.9.0)",
    "model-bakery (>=1.20.5,<2.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pypdf (>=6.0.0, <7.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "sqlglot (>=28.5.0,<29.0.0)",
    "social-auth-app-django (==5.7.0)",
    "django-guardian (>=2.4.0,<3.0.0)"
]

[tool.poetry]
package-mode = false